# back/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Gruzoviki-Yug API documentaion",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/",
            "description": "Грузовики-ЮГ dev server"
        },
        {
            "url": "https://api.грузовики-юг.рф/api/",
            "description": "Грузовики-ЮГ prod server"
        }
    ],
    "paths": {
        "/order/create": {
            "post": {
                "tags": [
                    "Cart"
                ],
                "summary": "Create an order",
                "operationId": "createOrder",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "total": {
                                        "type": "integer",
                                        "example": 100
                                    },
                                    "delivery": {
                                        "type": "string",
                                        "example": "Express"
                                    },
                                    "paymentMethod": {
                                        "type": "string",
                                        "example": "Credit Card"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "Pending"
                                    },
                                    "products": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer"
                                                },
                                                "quantity": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Заказ успешно создан"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "authorization": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}